+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## v1.3.0
 - #15 Availability zone
+- #13 Added support for both managed boot diagnostics, and bring your own storage account
+- #17 Support for setting timezones on Windows servers
 
 ## v1.2.0
 - #4 Add the possibility to create a network interface through the module