+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Changelog
 
 ## v1.3.0
-- #15 Availability zone
 - #13 Added support for both managed boot diagnostics, and bring your own storage account
+- #15 Availability zone
 - #17 Support for setting timezones on Windows servers
+- #18 Availability sets
 
 ## v1.2.0
 - #4 Add the possibility to create a network interface through the module